fix: make order type wire deterministic-serializable
Prior to this commit, order type information was encoded into a
`map[string]any`, which does not guarantee ordering neither on key access
nor on iteration. This is not the case in Python, so a mirror
implementation is not valid, enforcing the usage of structs which
provide deterministic order on serialization for both json and msgpack
codecs (if configured properly).